Added three table layouts
The backend database related file (where all of this is going) and the
user GUI interface and such. In the backend file, I need three tables.
The old books table, the new user table, and the loan table. I can use
the same format for all the tables, just like the books table. I then
need to link the loan table to the user table. Then I need to link the
user table to the main books table.